requirements.txt
boto3>=1.34
botocore>=1.34
python-dotenv>=1.0
datasets>=2.20
faiss-cpu>=1.8

langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# Choose ONE loader stack:

# (A) Docling (layout-aware)  —— RECOMMENDED
docling[all]>=1.7.0
pypdfium2>=4.30.0

# (B) Simple PDF fallback (comment these two above and use this if you skip Docling)
# pypdf>=4.2.0
langchain-docling
langchain_pinecone
pinecone
langchain_aws
ragas
bedrock_agentcore
bedrock-agentcore-starter-toolkit

langchain-pinecone>=0.1.0 
pinecone-client>=4.0.0